Modify the project structure
