+++ Makefile
@@ -4,7 +4,7 @@ OBJ = lexvec
 BUILD = go build --ldflags '-extldflags "-static"' -o $(OBJ)
 
 optimal:
-	CC="$(CC)" CGO_CFLAGS="$(CFLAGS) -march=native" $(BUILD)
+	cd cmd && CC="$(CC)" CGO_CFLAGS="$(CFLAGS) -march=native" $(BUILD) && cd -
 
 cross:
-	CC="$(CC)" CGO_CFLAGS="$(CFLAGS)" $(BUILD)
+	cd cmd && CC="$(CC)" CGO_CFLAGS="$(CFLAGS)" $(BUILD) && cd -
