+++ cmd/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"github.com/alexandres/lexvec"
+)
+
+func main() {
+	lexvec.Build()
+}
