+++ lexvec_test.go
@@ -1,4 +1,4 @@
-package main
+package lexvec
 
 import (
 	"os"
