+++ matrix.go
@@ -20,7 +20,7 @@
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package main
+package lexvec
 
 type matrix interface {
 	get(row, col idxUint) countUint
